+++ includes/admin/class.llms.admin.system-report.php
@@ -3,7 +3,7 @@
  * Admin System Report Class
  *
  * @since    2.1.0
- * @version  3.0.0
+ * @version  [version]
  */
 
 if ( ! defined( 'ABSPATH' ) ) { exit; }
@@ -14,7 +14,7 @@ class LLMS_Admin_System_Report {
 	 * Output the system report
 	 * @return   void
 	 * @since    2.1.0
-		 * @version  3.0.0
+	 * @version  3.0.0
 	 */
 	public static function output() {
 
@@ -38,7 +38,7 @@ class LLMS_Admin_System_Report {
 	/**
 	 * Output the copy for support box
 	 * @since    2.1.0
-		 * @version  3.0.0
+	 * @version  [version]
 	 */
 	public static function output_copy_box() {
 		?>
@@ -55,27 +55,29 @@ class LLMS_Admin_System_Report {
 		</div>
 		<script>
 			jQuery( document ).ready( function( $ ) {
-				var $textArea = $( '#llms-debug-report' ).find( 'textarea' );
+				var $textarea = $( '#llms-debug-report textarea' );
 
-				$(".llms-widget.settings-box").each( function( index, element ) {
-
-					var title = $(this).find('.llms-label').text();
+				$( '.llms-widget.settings-box' ).each( function( index, element ) {
+					var title = $( this ).find( '.llms-label' ).text();
 					title = title + '\n' + '-------------------------------------------';
-					var val = $(this).find('li').text().replace(/  /g, '').replace(/\t/g, '').replace(/\n\n/g, '\n');
-					$textArea.val($textArea.val() + title + '\n' + val + '\n\n');
-				});
+					var val = $( this ).find( 'li' ).text().replace(/  /g, '').replace(/\t/g, '').replace(/\n\n/g, '\n');
+					$textarea.val( $textarea.val() + title + '\n' + val + '\n\n' );
+				} );
 
-				$('#copy-for-support').on('click', function() {
-					$( '#llms-debug-report' ).find( 'textarea' ).show().select();
+				$( '#copy-for-support' ).on( 'click', function() {
+					$textarea.show().select();
 					try {
-						if(!document.execCommand('copy')) throw 'Not allowed.';
-					} catch(e) {
-						copyElement.remove();
-						console.log("document.execCommand('copy'); is not supported");
-						var text = $( '#debug-report' ).find( 'textarea' ).val();
-						prompt('Copy the text below. (ctrl c, enter)', text);
+						if ( ! document.execCommand( 'copy' ) ) {
+							throw 'Not allowed.';
+						}
+					} catch( e ) {
+						alert( 'copy the text below' );
 					}
-				})
+				} );
+
+				$textarea.on( 'click', function() {
+					$( this ).select();
+				} );
 			});
 		</script>
 		<?php
@@ -87,7 +89,7 @@ class LLMS_Admin_System_Report {
 	 * @param    arry     $data             array of data for the section
 	 * @return   void
 	 * @since    3.0.0
-	 * @version  3.0.0
+	 * @version  [version]
 	 */
 	public static function output_section( $section_title, $data ) {
 
@@ -96,8 +98,6 @@ class LLMS_Admin_System_Report {
 			$data = $data['active'];
 
 		}
-
-		// var_dump( $data );
 
 		?>
 		<div class="llms-widget-full top">
@@ -109,6 +109,8 @@ class LLMS_Admin_System_Report {
 							<li><p>
 							<?php if ( 'plugins' === $section_title ) : ?>
 								<?php self::plugin_item( $val ); ?>
+							<?php elseif ( 'template_overrides' === $section_title ) : ?>
+								<?php self::template_item( $val ); ?>
 							<?php else : ?>
 								<?php self::title( $key ); ?>: <strong><?php self::value( $val ); ?></strong>
 							<?php endif; ?>
@@ -135,6 +137,19 @@ class LLMS_Admin_System_Report {
 	}
 
 	/**
+	 * Output data related to an overridden template system report
+	 * @param    array     $data  array of template data
+	 * @return   void
+	 * @since    [version]
+	 * @version  [version]
+	 */
+	private static function template_item( $data ) {
+		echo '<strong>' . $data['template'] . ' (ver: ' . $data['core_version'] . ')</strong>: ';
+		echo '<code>' . $data['location'] . '</code> (ver: ' . $data['version'] . ')';
+	}
+
+
+	/**
 	 * Output the title for an item in the system report
 	 * @param    string     $key  title
 	 * @return   void
@@ -144,34 +159,18 @@ class LLMS_Admin_System_Report {
 	private static function title( $key ) {
 
 		$key = ucwords( str_replace( '_', ' ', $key ) );
-
-		// switch( $key ) {
-			// @todo allow translation here?
-		// }
-
 		echo $key;
 
 	}
 
 	/**
 	 * Output the value of an item in the system report
-	 * @todo  this should of might translate stuff one day?
 	 * @param    string     $val  value
 	 * @return   void
 	 * @since    3.0.0
 	 * @version  3.0.0
 	 */
 	private static function value( $val ) {
-
-		if ( is_array( $val ) ) {
-
-		}
-
-		// $val = ucwords( str_replace( '_', ' ', $val ) );
-
-		// switch( $val ) {
-			// @todo allow translation here?
-		// }
 
 		echo $val;