+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 == Changelog ==
 
+v3.11.2 - 2017/08/11
+--------------------
+
++ System status page
++ log viewer
++ template overrides on system report
+
+
 v3.11.1 - 2017/08/03
 --------------------